Create README.md
Tech. enthusiastic with some experience in data sciences and bioinformatics, and strong experience in Statistics and Geographic Information System (GIS). I am graduated from Biological Sciences and Plant Breeding and I found in ITs an awesome and interesting world with plenty of convergences with biological systems. To write a code and run it, is like a create a life form, which interacts with their environment, evolve and reproduce through pull-push. So, I feel passionate about being part of the evolution of these emergent life forms. which are codes and programs. 

Interested to collaborate with WebDev and Bioinformatic projects!

:)